Infrastructure, a few basic examples

+++ models/review_models.go
@@ -7,15 +7,15 @@ import (
 // TODO: add appropriate struct tags
 
 type Review struct {
-	id              uint32
+	id              uint64
 	rating          uint8
 	difficulty      uint8
 	interest        uint8
-	classId         uint32
+	classId         uint64
 	date            time.Time
 	anonymous       bool
 	text            string
 	professorName   string
-	helpfulCount    uint32
-	notHelpfulCount uint32
+	helpfulCount    uint64
+	notHelpfulCount uint64
 }
